requirements.txt
fastapi==0.121.1
httpx==0.28.1
pydantic==2.12.4
pydantic_settings==2.11.0
pyserial==3.5
python-dotenv==1.2.1
Requests==2.32.5
uvicorn==0.38.0
xmltodict==1.0.2
